# resources/api.spec.yaml
openapi: 3.0.0
info:
  description: This API lets users play the trivia game called Game Of Knowns
  version: 1.0.0
  title: Game Of Knowns - API
  # put the contact info for your development or API team
  contact:
    email: khili22.sharma@gmail.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# tags are used for organizing operations
...
paths:
...
  /winner:
    get:
      summary: Checks whether the player is the game of the winner or not.
      operationId: isWinner
      description: Displays the winner.
      parameters:
        - in: query
          name: gameId
          schema:
            type: string
          description: Unique identifier of the game which the player was playing
        - in: query
          name: playerId
          schema:
            type: string
          description: Unique identifier of the player
      responses:
        200:
          description: Winner object returns whether the player is winner or not
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Winner'
        404:
          description: Resource Not Found
        400:
          description: Bad input parameter
        401:
          description: Not authorized to access
components:
  schemas:
...
    Winner:
      type: object
      properties:
        winStatus:
          type: string
...
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/khilisharma/GameOfKnowns/1.0.0