# src/gen/java/io/swagger/model/Winner.java
/*
 * Game Of Knowns - API
 * This API lets users play the trivia game called Game Of Knowns
 *
 * OpenAPI spec version: 1.0.0
 * Contact: khili22.sharma@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.v3.oas.annotations.media.Schema;
import javax.validation.constraints.*;

/**
 * Winner
 */
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2020-07-16T04:40:34.487Z[GMT]")public class Winner   {
  @JsonProperty("winStatus")
  private String winStatus = null;

  public Winner winStatus(String winStatus) {
    this.winStatus = winStatus;
    return this;
  }

  /**
   * Get winStatus
   * @return winStatus
   **/
  @JsonProperty("winStatus")
  @Schema(description = "")
  public String getWinStatus() {
    return winStatus;
  }

  public void setWinStatus(String winStatus) {
    this.winStatus = winStatus;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Winner winner = (Winner) o;
    return Objects.equals(this.winStatus, winner.winStatus);
  }

  @Override
  public int hashCode() {
    return Objects.hash(winStatus);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Winner {\n");
    
    sb.append("    winStatus: ").append(toIndentedString(winStatus)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}